+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/serialization/CompressedStreamHandler.kt
@@ -7,11 +7,23 @@ import java.util.zip.GZIPInputStream
 object CompressedStreamHandler {
 
     fun wrapInput(input: InputStream): InputStream {
-        if (input.available() == 0) { return input }
+        if (input.available() == 2) { return input }
         var content = input
         if (!input.markSupported()) { content = BufferedInputStream(input) }
         content.mark(2)
-        val byteArray = content.readNBytes(2)
+
+        val firstValue = content.read()
+        if (firstValue == -1) {
+            content.reset()
+            return content
+        }
+        val secondValue = content.read()
+        if (secondValue == -1) {
+            content.reset()
+            return content
+        }
+        val byteArray = byteArrayOf(firstValue.toByte(), secondValue.toByte())
+
         content.reset()
         if (isGzipByteHeader(byteArray)) { return GZIPInputStream(content) }
         return content